# maquinaria_laravel/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API de Maquinaria",
        "description": "Documentación de la API de Mantenimientos y Maquinarias",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/categorias": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Listar todas las categorías",
                "operationId": "00d9b3c4359044674c4dafe7c158495d",
                "responses": {
                    "200": {
                        "description": "Lista de categorías con sus tipos de maquinaria",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/CategoriaMaquinaria"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Crear una nueva categoría",
                "operationId": "eb178d26b88f7e1e101a13507174ce37",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nombre"
                                ],
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Excavadoras"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Máquinas para excavación"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Categoría creada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoriaMaquinaria"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación"
                    }
                }
            }
        },
        "/api/categorias/{id}": {
            "get": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Obtener una categoría por ID",
                "operationId": "360086981d826f9bf4a8814e93d254ac",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoría encontrada",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoriaMaquinaria"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Categoría no encontrada"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Actualizar una categoría existente",
                "operationId": "d4e3da17f9c714cfd325c4df2ff1b91a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría a actualizar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nombre": {
                                        "type": "string",
                                        "example": "Grúas"
                                    },
                                    "descripcion": {
                                        "type": "string",
                                        "example": "Maquinaria de elevación"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Categoría actualizada exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CategoriaMaquinaria"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Categoría no encontrada"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categorias"
                ],
                "summary": "Eliminar una categoría",
                "operationId": "8c5a4ff8fb6c89dc732c9ecbac6df8ed",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la categoría a eliminar",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Categoría eliminada exitosamente"
                    },
                    "404": {
                        "description": "Categoría no encontrada"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "CategoriaMaquinaria": {
                "title": "Categoría de Maquinaria",
                "required": [
                    "nombre"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Excavadoras"
                    },
                    "descripcion": {
                        "type": "string",
                        "example": "Máquinas para excavar"
                    },
                    "tiposMaquinaria": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}